+++ vhdl_parser/src/contents.rs
@@ -39,12 +39,12 @@ impl Contents {
 
     #[cfg(test)]
     fn end(&self) -> Position {
-        let line = self.num_lines().saturating_sub(1) as u64;
+        let line = self.num_lines().saturating_sub(1) as u32;
         let character = self
             .lines
             .last()
             .map(|line| line.chars().map(|chr| chr.len_utf16()).sum())
-            .unwrap_or(0) as u64;
+            .unwrap_or(0) as u32;
         Position { line, character }
     }
